2012-10-15 13:55 UTC+0200 Viktor Szakats (harbour syenar.net)   * contrib/gtwvg/wvgwin.c   * contrib/hbamf/amfdec.c   * contrib/hbamf/amfenc.c   * contrib/hblzf/core.c   * contrib/hbpgsql/postgres.c   * contrib/hbxdiff/core.c   * contrib/hbzebra/codabar.c   * contrib/hbzebra/code39.c   * contrib/hbzebra/code93.c   * contrib/hbzebra/core.c   * contrib/hbzebra/qrcode.c   * contrib/rddads/ads1.c     ! fixed various 64-bit problems revealed by compiler warnings       (used casts, changed variables type, changed Harbour       API functions, please review especially hbamf, hbzebra and gtwvg)
  * extras/httpsrv/uhttpd.prg
    + changed to use keyboard events to detect window close,
      instead of deprecated and now dummy feature HB_GTI_NOTIFIERBLOCK

+++ harbour/contrib/hblzf/core.c
@@ -129,7 +129,7 @@ HB_FUNC( HB_LZF_COMPRESS )
 
          if( out_data )
          {
-            unsigned int uiResult = lzf_compress( in_data, in_len, out_data, out_len );
+            unsigned int uiResult = lzf_compress( in_data, ( unsigned int ) in_len, out_data, ( unsigned int ) out_len );
 
             if( uiResult != 0 )
             {
@@ -199,7 +199,7 @@ HB_FUNC( HB_LZF_DECOMPRESS )
                buffer_size <<= 1;
                buffer = ( char * ) hb_xrealloc( buffer, buffer_size + 1 );
 
-               uiResult = lzf_decompress( in_data, in_len, buffer, buffer_size );
+               uiResult = lzf_decompress( in_data, ( unsigned int ) in_len, buffer, ( unsigned int ) buffer_size );
             }
             while( uiResult == 0
 #if ! AVOID_ERRNO
